Support simple cases of string eval
Devel::Probe uses the CopFILE and CopLINE macros to decide whether perl
is interpreting a line of code which should trigger a probe. Typically
these macros have values like "my-cool-script.pl" and '42'.

However, those macros change if the flow of execution enters a string
eval: they instead return '(eval 1234)' and '2'. '1234' here is an
example of the internal perl ID for that eval, and '2' is the line
number offset to the start of the eval string -- _not_ the start of the
file.

This is problematic because puny humans cannot predict what identifier
perl will assign to a particular eval. This makes it more-or-less
impossible to set a probe for code loaded as part of a string eval.

Now, you might be thinking (as I did prior to this commit) that anybody
who uses string eval in production code is a menace to society, and does
not deserve nice tools like Devel::Probe. However, while testing
Devel::Probe, I found a program that uses string eval to hook 'require'
in a way that is transparent to all the code involved. This was pretty
compelling to me (and an important part of enabling Devel::Probe for
real users), so here we are.

This commit tries to jump past string eval frames and find their
more normal, non-string-eval parents. By reading CopFILE from an upper
stack frame, and combining the line offsets from each of the 'eval'
children, we can allow humans to set probes using line and filenames
just like their editor would indicate.

This particular implementation has a significant flaw: it can deal with
nested 'eval' without issue:

    eval '
        eval q|
            eval q^
                my $foo = "bar";
                my $bar = "baz";
            ^;
        |;
    ';

However, a sub declared & invoked inside an eval, like this:

    eval '
        sub my_cool_sub {
            my $foo = "bar";
            my $bar = "baz";
        }
        my_cool_sub();
    ';

... will still show '(eval 1234)' as the filename, and incorrect line
offsets. Not sure yet how to fix this; I suspect the answer might be to
hook OP_entereval and save line/file there... or something.